oss download, delete

+++ rust/src/common/oss/oss_dao.rs
@@ -1,4 +1,4 @@
-use anyhow::{Result, Ok};
+use anyhow::Result;
 use s3::request_trait::Request;
 use s3::{Region, Bucket, BucketConfiguration, creds::Credentials, command::Command};
 use s3::request::Reqwest;
@@ -7,6 +7,7 @@ pub type OssBucket = Bucket;
 
 #[derive(Debug)]
 pub struct StatObject {
+  pub last_modified: Option<String>,
   pub content_type: Option<String>,
   pub content_length: Option<i64>,
   pub filename: String,
@@ -65,12 +66,22 @@ pub async fn put_object<S: AsRef<str>>(
 
 pub async fn head_object(
   path: &str,
-) -> Result<StatObject> {
+) -> Result<Option<StatObject>> {
   let bucket = new_bucket()?;
-  let (res, _) = bucket.head_object(path).await?;
+  let res = bucket.head_object(path).await;
+  let res = match res {
+    Ok((res, _)) => res,
+    Err(e) => {
+      if e.to_string().contains("404") {
+        return Ok(None);
+      } else {
+        return Err(e.into());
+      }
+    }
+  };
   let filename: String = {
     if let Some(metadata) = res.metadata {
-      if let Some(f) = metadata.get("x-amz-meta-filename") {
+      if let Some(f) = metadata.get("filename") {
         f.to_owned()
       } else {
         "".to_owned()
@@ -80,12 +91,13 @@ pub async fn head_object(
     }
   };
   let stat = StatObject {
+    last_modified: res.last_modified,
     content_type: res.content_type,
     content_length: res.content_length,
     filename,
     etag: res.e_tag,
   };
-  Ok(stat)
+  Ok(stat.into())
 }
 
 pub async fn get_object(
@@ -99,6 +111,10 @@ pub async fn get_object(
 pub async fn delete_object(
   path: &str,
 ) -> Result<bool> {
+  let head = head_object(path).await?;
+  if head.is_none() {
+    return Ok(false);
+  }
   let bucket = new_bucket()?;
   bucket.delete_object(path).await?;
   Ok(true)
@@ -106,36 +122,12 @@ pub async fn delete_object(
 
 #[cfg(test)]
 mod test {
+  
   use anyhow::{Ok, Result};
 
-  use crate::common::oss::oss_dao::{new_bucket};
-  
   #[tokio::test]
   async fn test_oss() -> Result<()> {
-    
-    let mut bucket = new_bucket()?;
-    
-    let s3_path = "test.file";
-    let test = b"I'm going to S3!";
-    
-    // bucket.add_header("x-amz-meta-filename", urlencoding::encode("黄智勇.jpg").as_ref());
-
-    // let response_data = bucket.put_object_with_content_type(s3_path, test, "text/plain").await?;
-    // assert_eq!(response_data.status_code(), 200);
-
-    // let response_data = bucket.get_object(s3_path).await?;
-    // assert_eq!(response_data.status_code(), 200);
-    // assert_eq!(test, response_data.bytes());
-
-    let (head_object_result, code) = bucket.head_object(s3_path).await?;
-    assert_eq!(code, 200);
-    
-    println!("head_object_result.content_type: {:?}", head_object_result);
-    
-    println!("head_object_result.metadata: {:?}", head_object_result.metadata);
-    
-    // let response_data = bucket.delete_object(s3_path).await?;
-    // assert_eq!(response_data.status_code(), 204);
     Ok(())
   }
+  
 }
